requirements.txt
Pillow
torchvision
torch
numpy
matplotlib
tensorflow